message display complete with responses, but not updating with actions yet. deleted obsolete files
